+++ Bsp/Inc/motor.h
@@ -12,21 +12,21 @@ extern "C" {
 #include <stdint.h>
 
 typedef struct PidTypeDef {
-    float Kp;
-    float Ki;
-    float Kd;
-    float target;
-    float measure;
-    float err;
-    float err_last;
-    float integral;
-    float out_max;      // 输出限幅
-    float integral_max; // 积分限幅（抗饱和）
-    float out;
+    int32_t Kp;
+    int32_t Ki;
+    int32_t Kd;
+    int32_t target;
+    int32_t measure;
+    int32_t err;
+    int32_t err_last;
+    int32_t integral;
+    int32_t out_max;      // 输出限幅
+    int32_t integral_max; // 积分限幅（抗饱和）
+    int32_t out;
 } PidTypeDef;
 
-extern volatile float leftMotorDeg;
-extern volatile float rightMotorDeg;
+extern volatile int32_t leftMotorDeg;
+extern volatile int32_t rightMotorDeg;
 extern uint8_t pidEnabled;
 
 extern volatile int32_t leftMotorPwm;
@@ -36,8 +36,8 @@ void allMotorInit();
 void enablePid();
 void disablePid();
 
-void setLeftMotorDeg(float deg);
-void setRightMotorDeg(float deg);
+void setLeftMotorDeg(int32_t deg);
+void setRightMotorDeg(int32_t deg);
 
 void updateAllMotor();
 void updateRightMotorSpeed();
@@ -48,7 +48,7 @@ void leftMotorPid();
 
 void setLeftMotorPwm(int32_t pwm);
 void setRightMotorPwm(int32_t pwm);
-float getPidOutput(PidTypeDef* pid, float target, float measure);
+int32_t getPidOutput(PidTypeDef* pid, int32_t target, int32_t measure);
 
 #ifdef __cplusplus
 }
